+++ compiled/cdn-config/write.js
@@ -42,15 +42,17 @@ var __generator = (this && this.__generator) || function (thisArg, body) {
     }
 };
 exports.__esModule = true;
-exports.connectCdn = exports.setSmartcontract = void 0;
+exports.connectCdnByEnv = exports.connectCdn = exports.setHardhatSmartcontract = exports.incrementAbiConfiguration = exports.setAbi = exports.setAbiConfig = exports.setSmartcontract = void 0;
+var util_1 = require("./util");
 var read_1 = require("./read");
 var alicloud_1 = require("./alicloud");
+var hardhat_1 = require("./../utils/hardhat");
 var setSmartcontract = function (path, cdnClient, smartcontractPath, obj) { return __awaiter(void 0, void 0, void 0, function () {
     var idString, type, i, uploaded;
     return __generator(this, function (_a) {
         switch (_a.label) {
             case 0:
-                if (global.config === undefined || global.config === null) {
+                if (global.seascapeCdnConfig === undefined || global.seascapeCdnConfig === null) {
                     console.log({
                         error_path: 'src/cdn-config/write.setSmartcontract',
                         line: 'no_config',
@@ -60,20 +62,20 @@ var setSmartcontract = function (path, cdnClient, smartcontractPath, obj) { retu
                 }
                 idString = smartcontractPath.networkId.toString();
                 type = smartcontractPath.type;
-                if (!global.config[idString]) {
-                    global.config[idString] = {};
+                if (!global.seascapeCdnConfig[idString]) {
+                    global.seascapeCdnConfig[idString] = {};
                 }
-                if (!global.config[idString][type]) {
-                    global.config[idString][type] = [];
+                if (!global.seascapeCdnConfig[idString][type]) {
+                    global.seascapeCdnConfig[idString][type] = [];
                 }
                 i = (0, read_1.contractIndex)(idString, type, obj.name);
-                if (!i) {
-                    global.config[idString][type].push(obj);
+                if (i === false) {
+                    global.seascapeCdnConfig[idString][type].push(obj);
                 }
                 else {
-                    global.config[idString][type][i] = obj;
+                    global.seascapeCdnConfig[idString][type][i] = obj;
                 }
-                return [4 /*yield*/, (0, alicloud_1.uploadConfig)(path, cdnClient, global.config)];
+                return [4 /*yield*/, (0, alicloud_1.uploadConfig)(path, cdnClient, global.seascapeCdnConfig)];
             case 1:
                 uploaded = _a.sent();
                 if (uploaded === null) {
@@ -89,5 +91,130 @@ var setSmartcontract = function (path, cdnClient, smartcontractPath, obj) { retu
     });
 }); };
 exports.setSmartcontract = setSmartcontract;
+var setAbiConfig = function (cdnClient, smartcontractName, abiConfig) { return __awaiter(void 0, void 0, void 0, function () {
+    var uploaded;
+    return __generator(this, function (_a) {
+        switch (_a.label) {
+            case 0: return [4 /*yield*/, (0, alicloud_1.uploadAbiConfig)(cdnClient, smartcontractName, abiConfig)];
+            case 1:
+                uploaded = _a.sent();
+                if (uploaded === null) {
+                    console.log({
+                        error_path: 'src/cdn-config/write.setAbiConfig',
+                        line: 'no_upload',
+                        message: "Please fix the Alicloud credentials"
+                    });
+                    return [2 /*return*/, false];
+                }
+                return [2 /*return*/, true];
+        }
+    });
+}); };
+exports.setAbiConfig = setAbiConfig;
+var setAbi = function (cdnClient, smartcontractName, abiConfig, abi) { return __awaiter(void 0, void 0, void 0, function () {
+    var uploaded;
+    return __generator(this, function (_a) {
+        switch (_a.label) {
+            case 0: return [4 /*yield*/, (0, alicloud_1.uploadAbi)(cdnClient, smartcontractName, abiConfig, abi)];
+            case 1:
+                uploaded = _a.sent();
+                if (uploaded === null) {
+                    console.log({
+                        error_path: 'src/cdn-config/write.setAbi',
+                        line: 'no_upload',
+                        message: "Please fix the Alicloud credentials"
+                    });
+                    return [2 /*return*/, false];
+                }
+                return [2 /*return*/, true];
+        }
+    });
+}); };
+exports.setAbi = setAbi;
+var incrementAbiConfiguration = function (client, smartcontractName) { return __awaiter(void 0, void 0, void 0, function () {
+    var abiConfig, abiUpdated;
+    return __generator(this, function (_a) {
+        switch (_a.label) {
+            case 0: return [4 /*yield*/, (0, read_1.abiConfig)(smartcontractName)];
+            case 1:
+                abiConfig = _a.sent();
+                console.log("Before incremention");
+                console.log(abiConfig);
+                abiConfig.version++;
+                return [4 /*yield*/, (0, exports.setAbiConfig)(client, smartcontractName, abiConfig)];
+            case 2:
+                abiUpdated = _a.sent();
+                if (!abiUpdated) {
+                    console.error("Abi config wasn't updated.");
+                    return [2 /*return*/, false];
+                }
+                console.log("After incremention");
+                console.log(abiConfig);
+                return [2 /*return*/, abiConfig];
+        }
+    });
+}); };
+exports.incrementAbiConfiguration = incrementAbiConfiguration;
+/**
+ *
+ * @param params Object containing the following parameters
+ */
+var setHardhatSmartcontract = function (params) { return __awaiter(void 0, void 0, void 0, function () {
+    var client, abiConfig, abi, abiSetted, path, smartcontractPath, initialized, smartcontract, updated;
+    return __generator(this, function (_a) {
+        switch (_a.label) {
+            case 0: return [4 /*yield*/, (0, alicloud_1.connectionByEnvironment)()];
+            case 1:
+                client = _a.sent();
+                if (client === undefined) {
+                    return [2 /*return*/, false];
+                }
+                return [4 /*yield*/, (0, exports.incrementAbiConfiguration)(client, params.contractName)];
+            case 2:
+                abiConfig = _a.sent();
+                console.log('after the function');
+                console.log(abiConfig);
+                return [4 /*yield*/, (0, hardhat_1.abiFile)(params.contractName)];
+            case 3:
+                abi = _a.sent();
+                if (!abi) {
+                    console.log("Failed to load the abi of ".concat(params.contractName, " in hardhat framework. Please upload manually as ").concat(abiConfig.version));
+                    return [2 /*return*/, false];
+                }
+                return [4 /*yield*/, (0, exports.setAbi)(client, params.contractName, abiConfig, abi)];
+            case 4:
+                abiSetted = _a.sent();
+                if (!abiSetted) {
+                    return [2 /*return*/, false];
+                }
+                path = { project: params.projectName, env: params.projectEnv };
+                smartcontractPath = { networkId: params.networkId, type: params.contractType };
+                console.log("The cdn list path where smartcontract object will be");
+                console.log(smartcontractPath);
+                return [4 /*yield*/, (0, read_1.initConfig)(path)];
+            case 5:
+                initialized = _a.sent();
+                if (!initialized) {
+                    console.log("Global initializiation failed");
+                    process.exit(1);
+                }
+                smartcontract = {
+                    name: params.contractName,
+                    address: params.deployedInstance.address,
+                    txid: params.deployedInstance.deployTransaction.hash,
+                    abi: (0, util_1.cdnAbiUrl)(params.contractName, abiConfig, true)
+                };
+                console.log("The smartcontract object in the cdn config is");
+                console.log(smartcontract);
+                return [4 /*yield*/, (0, exports.setSmartcontract)(path, client, smartcontractPath, smartcontract)];
+            case 6:
+                updated = _a.sent();
+                console.log("Was CDN updated successfully? ".concat(updated));
+                return [2 /*return*/, updated];
+        }
+    });
+}); };
+exports.setHardhatSmartcontract = setHardhatSmartcontract;
 exports.connectCdn = alicloud_1.connection;
+exports.connectCdnByEnv = alicloud_1.connectionByEnvironment;
 //# sourceMappingURL=write.js.map