+++ src/cdn-config/util.ts
@@ -18,14 +18,49 @@ export interface SmartcontractPath {
     type: string
 }
 
+export interface HardhatSmartcontractConfig {
+    networkId?: number,
+    projectName: string,
+    projectEnv: string,
+    contractType: string, 
+    contractName: string, 
+    deployedInstance: any
+}
+
+export interface AbiConfig {
+    version: number
+}
+
 export const SEASCAPE_CDN = 'https://cdn.seascape.network/';
 
 export let cdnConfigUrl = (path: ConfigPath): string => {
     return `${SEASCAPE_CDN}${path.project}/${path.env}/config.json`;
 }
 
+export const cdnAbiConfigUrl = (contractName: string, fullAddress = true) : string => {
+    if (fullAddress) {
+        return `${SEASCAPE_CDN}abi/${contractName}/info.json`;
+    } else {
+        return `/abi/${contractName}/info.json`;
+    }
+}
+
+export const cdnAbiUrl = (contractName: string, config: AbiConfig, fullAddress = true): string => {
+    if (fullAddress) {
+        return `${SEASCAPE_CDN}abi/${contractName}/${config.version.toString()}.json`;
+    } else {
+        return `/abi/${contractName}/${config.version.toString()}.json`;
+    }
+}
+
+export let defaultAbiConfig = () : AbiConfig => {
+    return {
+        version: 0
+    } as AbiConfig;
+}
+
 export let validateConfNetwork = (networkId: string): boolean => {
-    if ((global as any).config === undefined || (global as any).config === null) {
+    if ((global as any).seascapeCdnConfig === undefined || (global as any).seascapeCdnConfig === null) {
         console.log({
             error_path: 'src/utils/config.validateConfNetwork',
             line: 'no_config',
@@ -34,7 +69,7 @@ export let validateConfNetwork = (networkId: string): boolean => {
         return false;
     }
 
-    if ((global as any).config[networkId] === undefined || (global as any).config[networkId] === null) {
+    if ((global as any).seascapeCdnConfig[networkId] === undefined || (global as any).seascapeCdnConfig[networkId] === null) {
         console.log({
             error_path: 'src/utils/config.validateConfNetwork',
             line: 'no_network_id',