+++ test/read-abi-config.ts
@@ -1,0 +1,6 @@
+import { CdnRead, ConfigPath } from "../src/index";
+
+(async () => {
+    let abiConfig = await CdnRead.abiConfig('Greeter');
+    console.log(abiConfig);
+})();