+++ test/read-abi.ts
@@ -1,0 +1,7 @@
+import { CdnRead } from "../src/index";
+
+(async () => {
+    let abiConfig = await CdnRead.abiConfig('Greeter');
+    let abi = await CdnRead.abi('Greeter', abiConfig);
+    console.log(abi);
+})();